update v2 js

+++ src/js/03-feedback.js
@@ -3,24 +3,29 @@ import throttle from 'lodash.throttle';
 const input = document.querySelector('input');
 const message = document.querySelector('textarea');
 const submitForm = document.querySelector('form');
-
-const inputChanged = event => {
-  localStorage.setItem(
-    'feedback-form-state',
-    JSON.stringify({ email: event.currentTarget.value, message: message.value })
-  );
+const formFeedback = {
+  email: '',
+  message: '',
 };
-const messageChanged = event => {
-  localStorage.setItem(
-    'feedback-form-state',
-    JSON.stringify({ email: input.value, message: event.currentTarget.value })
-  );
+const formChanged = event => {
+  formFeedback.email = input.value;
+  formFeedback.message = message.value;
+  localStorage.setItem('feedback-form-state', JSON.stringify(formFeedback));
 };
 
 const submitBtnClick = event => {
-  localStorage.clear();
-  console.log(JSON.parse(localStorage.getItem('feedback-form-state')));
+  event.preventDefault();
+  const formElements = event.currentTarget.elements;
+  const formData = new FormData(event.currentTarget);
+  if (formElements.email.value === '' || formElements.message.value === '') {
+    alert('Всі поля повинні бути не пусті!!!');
+  } else {
+    console.log(JSON.parse(localStorage.getItem('feedback-form-state')));
+    event.currentTarget.reset();
+    localStorage.clear();
+  }
 };
+
 const updateBrowserform = () => {
   formElements = submitForm.elements;
   try {
@@ -37,6 +42,6 @@ const updateBrowserform = () => {
   } catch (error) {}
 };
 updateBrowserform();
-input.addEventListener('input', throttle(inputChanged, 500));
-message.addEventListener('input', throttle(messageChanged, 500));
+input.addEventListener('input', throttle(formChanged, 500));
+message.addEventListener('input', throttle(formChanged, 500));
 submitForm.addEventListener('submit', submitBtnClick);
